app/requirements.txt
anyio==3.6.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.0.4
execnet==1.9.0
fastapi==0.88.0
fastapi-cache2==0.1.8
h11==0.14.0
httpcore==0.16.3
httpx==0.23.1
icmplib==3.0.3
idna==3.4
ifcfg==0.23
iniconfig==1.1.1
packaging==22.0
pendulum==2.1.2
pluggy==1.0.0
port-range==2.2.0
pydantic==1.10.2
pytest==7.2.0
pytest-xdist==3.1.0
python-dateutil==2.8.2
pytzdata==2020.1
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
